requirements.txt
selenium
pytest
requests
allure-pytest
pytest-rerunfailures
jsonpath
pyyaml